+++ include/audio_api/porcelain.h
@@ -16,7 +16,9 @@ RECOMP_IMPORT("magemods_audio_api", bool AudioApi_AddAudioFileFromFs(AudioApiFil
 RECOMP_IMPORT("magemods_audio_api", uintptr_t AudioApi_GetResourceDevAddr(u32 resourceId, u32 arg1, u32 arg2));
 
 RECOMP_IMPORT("magemods_audio_api", s32 AudioApi_CreateStreamedSequence(AudioApiFileInfo* info, AudioApiSequenceIO seqIO));
+RECOMP_IMPORT("magemods_audio_api", s32 AudioApi_CreateStreamedBgmEx(AudioApiFileInfo2* info2, char* dir, char* filename, AudioApiSequenceIO seqIO));
 RECOMP_IMPORT("magemods_audio_api", s32 AudioApi_CreateStreamedBgm(AudioApiFileInfo* info, char* dir, char* filename, AudioApiSequenceIO seqIO));
+RECOMP_IMPORT("magemods_audio_api", s32 AudioApi_CreateStreamedFanfareEx(AudioApiFileInfo2* info2, char* dir, char* filename, AudioApiSequenceIO seqIO));
 RECOMP_IMPORT("magemods_audio_api", s32 AudioApi_CreateStreamedFanfare(AudioApiFileInfo* info, char* dir, char* filename, AudioApiSequenceIO seqIO));
 
 #endif