+++ CHANGELOG.md
@@ -5,12 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.8.0] - 2026-02-26
+## [0.8.1] - 2026-02-26
+### Added
+- AudioApiFileInfo2, AudioApi_CreateStreamedBgmEx, AudioApi_CreateStreamedFanfareEx, allowing to define a volume offset per track
 ### Fixed
 - Main bgm not correctly unmuted after enemy cleared
 - Note priority, which caused N64 stealing soundtrack notes whenever it felt like it
 ### Changed
 - better concern separation to make it easier to control audio behavior without locking down API to a specific use case
+- Streamed audio dropping left channel due to note stealing prio and max notes restrictions
+- SFX not working during Final Hours after seperation of concerns patch
 
 ## [0.7.3] - 2026-02-23
 ### Fixed